# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PDC Darts Career Mode API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SetNameRequest {
  name: string;
}

export interface BuyEquipmentRequest {
  id: string;
}

export interface MatchResult {
  legs_won: number;
  legs_lost: number;
  my_avg?: number;
  my_180s?: number;
  my_hf?: number;
  my_co_pct?: number;
}

export interface RoundInfo {
  name: string;
  first_to: number;
  format: string;
}

export interface Matchup {
  player1: string;
  player2: string;
}

export interface OomEntry {
  name: string;
  geld: number;
}

export interface Achievement {
  name: string;
  desc: string;
  unlocked: boolean;
}

export interface Headline {
  titel: string;
  text: string;
}

export interface Sponsor {
  name: string;
  ziel_typ: string;
  ziel_wert: number;
  aktuell: number;
  turniere_zeit: number;
  belohnung: number;
  text: string;
}

export interface TournamentRecord {
  name: string;
  typ: string;
  runde: string;
  ergebnis: string;
  preisgeld: number;
  saison: number;
  avg: number;
}

export interface RankingRecord {
  platz: number;
  saison: number;
  turnier: string;
}

export interface EquipmentItem {
  id: string;
  name: string;
  beschreibung: string;
  preis: number;
  kategorie: string;
  bonus_typ: string;
  bonus_wert: number;
  owned: boolean;
}

export interface CalendarEntry {
  index: number;
  name: string;
  typ: string;
  format: string;
  min_platz?: number | null;
  status: string;
  qualifiziert: boolean;
}

export interface H2HRecord {
  name: string;
  siege: number;
  niederlagen: number;
}

export interface GegnerForm {
  form: string;
  bonus: number;
}

export type CareerStateAchievements = { [key: string]: Achievement };

export interface CareerState {
  spieler_name: string;
  name_set: boolean;
  hat_tourcard: boolean;
  q_school_punkte: number;
  order_of_merit_geld: number;
  bank_konto: number;
  saison_jahr: number;
  turnier_laeuft: boolean;
  aktuelle_runde: number;
  gegner_name: string;
  gegner_avg: number;
  gegner_form: GegnerForm;
  stats_spiele: number;
  stats_siege: number;
  stats_legs_won: number;
  stats_legs_lost: number;
  stats_180s: number;
  stats_highest_finish: number;
  achievements: CareerStateAchievements;
  aktiver_sponsor?: Sponsor | null;
  letzte_schlagzeile?: Headline | null;
  platz: number;
  turnier_name: string;
  quote: number;
  gesamt_avg: number;
  gesamt_co: number;
  oom: OomEntry[];
  runden_info: RoundInfo;
  matchups: Matchup[];
  h2h_siege: number;
  h2h_niederlagen: number;
  walk_on_video?: string | null;
  ranking_verlauf: RankingRecord[];
  avg_bonus: number;
  checkout_bonus: number;
}

export interface CareerResponse {
  career: CareerState;
  messages: string[];
}

export interface TournamentHistoryResponse {
  history: TournamentRecord[];
}

export interface H2HResponse {
  records: H2HRecord[];
}

export interface CalendarResponse {
  entries: CalendarEntry[];
  aktuelles_turnier_index: number;
}

export interface EquipmentResponse {
  items: EquipmentItem[];
  bank_konto: number;
}