# lib/api-zod/src/generated/types/careerState.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PDC Darts Career Mode API
 * OpenAPI spec version: 0.1.0
 */
import type { CareerStateAchievements } from "./careerStateAchievements";
import type { GegnerForm } from "./gegnerForm";
import type { Headline } from "./headline";
import type { Matchup } from "./matchup";
import type { OomEntry } from "./oomEntry";
import type { RankingRecord } from "./rankingRecord";
import type { RoundInfo } from "./roundInfo";
import type { Sponsor } from "./sponsor";

export interface CareerState {
  spieler_name: string;
  name_set: boolean;
  hat_tourcard: boolean;
  q_school_punkte: number;
  order_of_merit_geld: number;
  bank_konto: number;
  saison_jahr: number;
  turnier_laeuft: boolean;
  aktuelle_runde: number;
  gegner_name: string;
  gegner_avg: number;
  gegner_form: GegnerForm;
  stats_spiele: number;
  stats_siege: number;
  stats_legs_won: number;
  stats_legs_lost: number;
  stats_180s: number;
  stats_highest_finish: number;
  achievements: CareerStateAchievements;
  aktiver_sponsor?: Sponsor | null;
  letzte_schlagzeile?: Headline | null;
  platz: number;
  turnier_name: string;
  quote: number;
  gesamt_avg: number;
  gesamt_co: number;
  oom: OomEntry[];
  runden_info: RoundInfo;
  matchups: Matchup[];
  h2h_siege: number;
  h2h_niederlagen: number;
  walk_on_video?: string | null;
  ranking_verlauf: RankingRecord[];
  avg_bonus: number;
  checkout_bonus: number;
}