# lib/api-zod/src/generated/types/gegnerForm.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PDC Darts Career Mode API
 * OpenAPI spec version: 0.1.0
 */

export interface GegnerForm {
  form: string;
  bonus: number;
}