# lib/api-zod/src/generated/types/rankingRecord.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PDC Darts Career Mode API
 * OpenAPI spec version: 0.1.0
 */

export interface RankingRecord {
  platz: number;
  saison: number;
  turnier: string;
}